fix: connector not connected error
